refactoring the tests for the backend to be reusable when testing other backend classes
