+++ grid2op/tests/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ["BaseBackendTest"]
+
+# import grid2op.tests.BaseBackendTest
